Work on fixing Zeus Modules
