+++ addons/danger/CfgVehicles.hpp
@@ -15,6 +15,7 @@ class CfgVehicles {
     };
 
     class Module_F;
+
     class GVAR(SetRadio) : Module_F {
         author = "LAMBS Dev Team";
         _generalMacro = QGVAR(SetRadio);
@@ -22,10 +23,13 @@ class CfgVehicles {
         scopeCurator = 2;
         displayName = "Configure Long-range Radio";
         category = "Lambs_Danger_Cat";
-        function = QFUNC(moduleSetRadio);
-        functionPriority = 1;
         icon = "\A3\ui_f\data\igui\cfg\simpleTasks\types\intel_ca.paa";
-        isGlobal = 0;
+        class EventHandlers {
+            class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers_base {};
+            class ADDON {
+                init = QUOTE([ARR_3(_this,true,true)] call FUNC(moduleSetRadio));
+            };
+        };
     };
 
     class GVAR(DisableAI) : Module_F {
@@ -35,22 +39,27 @@ class CfgVehicles {
         scopeCurator = 2;
         displayName = "Disable Unit AI";
         category = "Lambs_Danger_Cat";
-        function = QFUNC(moduleDisableAI);
-        functionPriority = 1;
         icon = "\A3\ui_f\data\igui\cfg\simpleTasks\types\intel_ca.paa";
-        isGlobal = 0;
+        class EventHandlers {
+            class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers_base {};
+            class ADDON {
+                init = QUOTE([ARR_3(_this,true,true)] call FUNC(moduleDisableAI));
+            };
+        };
     };
 
     class GVAR(DisableGroupAI) : Module_F {
-        author = "LAMBS Dev Team";
         _generalMacro = QGVAR(DisableGroupAI);
         scope = 1;
         scopeCurator = 2;
         displayName = "Configure Group AI";
         category = "Lambs_Danger_Cat";
-        function = QFUNC(moduleDisableGroupAI);
-        functionPriority = 1;
         icon = "\A3\ui_f\data\igui\cfg\simpleTasks\types\intel_ca.paa";
-        isGlobal = 0;
+        class EventHandlers {
+            class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers_base {};
+            class ADDON {
+                init = QUOTE([ARR_3(_this,true,true)] call FUNC(moduleDisableGroupAI));
+            };
+        };
     };
 };
