+++ addons/main/CfgFactionClasses.hpp
@@ -2,11 +2,17 @@ class CfgFactionClasses {
     class NO_CATEGORY;
     class Lambs_Danger_Cat: NO_CATEGORY {
         displayName = "LAMBS Danger FSM";
+        priority = 2;
+        side = 7;
     };
     class Lambs_Danger_WP_Cat: NO_CATEGORY {
         displayName = "LAMBS Danger Waypoints";
+        priority = 2;
+        side = 7;
     };
     class Lambs_Danger_WP_Search_Cat: NO_CATEGORY {
         displayName = "LAMBS Danger Search Modules";
+        priority = 2;
+        side = 7;
     };
 };
